Learn classifier layers on top of a constituency parser as a new classifier model.
Includes a flag to backprop into the constituency parser's weights.
Makes the classifier much more accurate (and slower to train)

Also test saving & loading a model
Add a basic test of the forward pass for the constituency classifier.

In the constituency classifier, extract one layer down instead of ROOT as an option